Create Generates a random password (JavaScript)
This JavaScript code defines a generatePassword function that creates a random password of a specified length. It utilizes a set of characters (lowercase letters, uppercase letters, numbers, and symbols) stored in the charset variable. The function iterates through the charset to randomly select characters and concatenate them to form the password. Adjust the passwordLength variable to set the desired length of the generated password. When executed, it prints the generated password to the console. This code can be used as a utility to generate secure passwords for various applications.